Include all faest variants - no masking

+++ crypto_sign/faest_em_128f/ref/NIST-KATs/PQCgenKAT_sign.c
@@ -50,7 +50,7 @@ int main() {
   unsigned char msg[3300];
   unsigned char entropy_input[48];
   unsigned char *m, *sm, *m1;
-  unsigned long long mlen, smlen, mlen1;
+  unsigned long mlen, smlen, mlen1;
   int count;
   int done;
   unsigned char pk[CRYPTO_PUBLICKEYBYTES], sk[CRYPTO_SECRETKEYBYTES];
@@ -77,7 +77,7 @@ int main() {
     randombytes(seed, 48);
     fprintBstr(fp_req, "seed = ", seed, 48);
     mlen = 33 * (i + 1);
-    fprintf(fp_req, "mlen = %llu\n", mlen);
+    fprintf(fp_req, "mlen = %lu\n", mlen);
     randombytes(msg, mlen);
     fprintBstr(fp_req, "msg = ", msg, mlen);
     fprintf(fp_req, "pk =\n");
@@ -113,12 +113,12 @@ int main() {
     randombytes_init(seed, NULL, 256);
 
     if (FindMarker(fp_req, "mlen = "))
-      fscanf(fp_req, "%llu", &mlen);
+      fscanf(fp_req, "%lu", &mlen);
     else {
       printf("ERROR: unable to read 'mlen' from <%s>\n", fn_req);
       return KAT_DATA_ERROR;
     }
-    fprintf(fp_rsp, "mlen = %llu\n", mlen);
+    fprintf(fp_rsp, "mlen = %lu\n", mlen);
 
     m  = (unsigned char*)calloc(mlen, sizeof(unsigned char));
     m1 = (unsigned char*)calloc(mlen + CRYPTO_BYTES, sizeof(unsigned char));
@@ -142,7 +142,7 @@ int main() {
       printf("crypto_sign returned <%d>\n", ret_val);
       return KAT_CRYPTO_FAILURE;
     }
-    fprintf(fp_rsp, "smlen = %llu\n", smlen);
+    fprintf(fp_rsp, "smlen = %lu\n", smlen);
     fprintBstr(fp_rsp, "sm = ", sm, smlen);
     fprintf(fp_rsp, "\n");
 
@@ -152,7 +152,7 @@ int main() {
     }
 
     if (mlen != mlen1) {
-      printf("crypto_sign_open returned bad 'mlen': Got <%llu>, expected <%llu>\n", mlen1, mlen);
+      printf("crypto_sign_open returned bad 'mlen': Got <%lu>, expected <%lu>\n", mlen1, mlen);
       return KAT_CRYPTO_FAILURE;
     }
